[TS-289] Refactor TR service
