+++ libs/application/template-api-modules/src/lib/modules/templates/index.ts
@@ -94,8 +94,8 @@ import { CitizenshipService } from './directorate-of-immigration/citizenship/cit
 
 import { DrivingLicenseDuplicateModule } from './driving-license-duplicate/driving-license-duplicate.module'
 import { DrivingLicenseDuplicateService } from './driving-license-duplicate/driving-license-duplicate.service'
-import { OldAgePensionModule } from './social-insurance-administration/social-insurance-administration.module'
-import { OldAgePensionService } from './social-insurance-administration/social-insurance-administration.service'
+import { SocialInsuranceAdministrationModule } from './social-insurance-administration/social-insurance-administration.module'
+import { SocialInsuranceAdministrationService } from './social-insurance-administration/social-insurance-administration.service'
 
 export const modules = [
   ReferenceTemplateModule,
@@ -146,7 +146,7 @@ export const modules = [
   DrivingLicenseBookUpdateInstructorModule,
   DrivingLearnersPermitModule,
   DrivingLicenseDuplicateModule,
-  OldAgePensionModule,
+  SocialInsuranceAdministrationModule,
   CitizenshipModule,
 ]
 
@@ -199,6 +199,6 @@ export const services = [
   DrivingLicenseBookUpdateInstructorService,
   DrivingLearnersPermitService,
   DrivingLicenseDuplicateService,
-  OldAgePensionService,
+  SocialInsuranceAdministrationService,
   CitizenshipService,
 ]
