+++ libs/application/templates/social-insurance-administration/core/.babelrc
@@ -1,0 +1,4 @@
+{
+  "presets": ["@nx/react/babel"],
+  "plugins": []
+}
