+++ libs/application/templates/social-insurance-administration/core/.eslintrc.json
@@ -1,0 +1,10 @@
+{
+  "extends": ["plugin:@nx/react", "../../../../../.eslintrc.json"],
+  "ignorePatterns": ["!**/*"],
+  "rules": {},
+  "overrides": [
+    { "files": ["*.ts", "*.tsx", "*.js", "*.jsx"], "rules": {} },
+    { "files": ["*.ts", "*.tsx"], "rules": {} },
+    { "files": ["*.js", "*.jsx"], "rules": {} }
+  ]
+}
