+++ libs/application/templates/social-insurance-administration/core/README.md
@@ -1,0 +1,3 @@
+# Social Insurance Administration Core
+
+This is an core library for Social Insurance Administration applications that holds shared components, utils, types etc.
