+++ libs/application/templates/social-insurance-administration/core/jest.config.js
@@ -1,0 +1,12 @@
+module.exports = {
+  displayName: 'application-templates-social-insurance-administration-core',
+  preset: './jest.preset.js',
+  rootDir: '../../../../..',
+  roots: [__dirname],
+  transform: {
+    '^.+\\.[tj]sx?$': ['babel-jest', { presets: ['@nx/react/babel'] }],
+  },
+  moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx'],
+  coverageDirectory:
+    '<rootDir>/coverage/libs/application/templates/social-insurance-administration/core',
+}
