+++ libs/application/templates/social-insurance-administration/core/src/types.ts
@@ -1,0 +1,16 @@
+export interface BankInfo {
+  bank?: string
+  ledger?: string
+  accountNumber?: string
+  iban?: string
+  swift?: string
+  foreignBankName?: string
+  foreignBankAddress?: string
+  currency?: string
+}
+
+export interface ApplicantInfo {
+  bankAccount?: BankInfo
+  phoneNumber?: string
+  emailAddress?: string
+}
