+++ libs/application/templates/social-insurance-administration/old-age-pension/src/lib/oldAgePensionUtils.spec.ts
@@ -17,12 +17,7 @@ import {
   getApplicationExternalData,
   isEarlyRetirement,
   filterValidEmployers,
-  friendlyFormatSWIFT,
-  getBankIsk,
   shouldNotUpdateBankAccount,
-  friendlyFormatIBAN,
-  validIBAN,
-  validSWIFT,
 } from './oldAgePensionUtils'
 import { ApplicationType, MONTHS } from './constants'
 import * as kennitala from 'kennitala'
@@ -315,31 +310,6 @@ describe('filterValidEmployers', () => {
   })
 })
 
-describe('getBankIsk', () => {
-  it('should return icelandic bank number if bank, ledger and account number is returned', () => {
-    const application = buildApplication({
-      externalData: {
-        socialInsuranceAdministrationApplicant: {
-          data: {
-            bankAccount: {
-              bank: '2222',
-              ledger: '00',
-              accountNumber: '123456',
-            },
-          },
-          date: new Date(),
-          status: 'success',
-        },
-      },
-    })
-
-    const { bankInfo } = getApplicationExternalData(application.externalData)
-    const bankNumer = getBankIsk(bankInfo)
-
-    expect('222200123456').toEqual(bankNumer)
-  })
-})
-
 describe('shouldNotUpdateBankAccount', () => {
   it('should return true if bank account returned from TR is not changed', () => {
     const application = buildApplication({
@@ -405,157 +375,5 @@ describe('shouldNotUpdateBankAccount', () => {
     )
 
     expect(false).toEqual(res)
-  })
-})
-
-describe('friendlyFormat & valid', () => {
-  it('format swift', () => {
-    const application = buildApplication({
-      externalData: {
-        socialInsuranceAdministrationApplicant: {
-          data: {
-            bankAccount: {
-              iban: 'NL91ABNA0417164300',
-              swift: 'NEDSZAJJXXX',
-              foreignBankName: 'Heiti banka',
-              foreignBankAddress: 'Heimili banka',
-              currency: 'EUR',
-            },
-          },
-          date: new Date(),
-          status: 'success',
-        },
-      },
-    })
-
-    const { bankInfo } = getApplicationExternalData(application.externalData)
-    const formattedSWIFT = friendlyFormatSWIFT(bankInfo.swift)
-
-    expect('NEDS ZA JJ XXX').toEqual(formattedSWIFT)
-  })
-
-  it('format iban', () => {
-    const application = buildApplication({
-      externalData: {
-        socialInsuranceAdministrationApplicant: {
-          data: {
-            bankAccount: {
-              iban: 'NL91ABNA0417164300',
-              swift: 'NEDSZAJJXXX',
-              foreignBankName: 'Heiti banka',
-              foreignBankAddress: 'Heimili banka',
-              currency: 'EUR',
-            },
-          },
-          date: new Date(),
-          status: 'success',
-        },
-      },
-    })
-
-    const { bankInfo } = getApplicationExternalData(application.externalData)
-    const formattedIBAN = friendlyFormatIBAN(bankInfo.iban)
-
-    expect('NL91 ABNA 0417 1643 00').toEqual(formattedIBAN)
-  })
-
-  it('valid iban - should return false because the check digits should be numbers', () => {
-    const application = buildApplication({
-      externalData: {
-        socialInsuranceAdministrationApplicant: {
-          data: {
-            bankAccount: {
-              iban: 'NLLLABNA0417164300',
-              swift: 'NEDSZAJJXXX',
-              foreignBankName: 'Heiti banka',
-              foreignBankAddress: 'Heimili banka',
-              currency: 'EUR',
-            },
-          },
-          date: new Date(),
-          status: 'success',
-        },
-      },
-    })
-
-    const { bankInfo } = getApplicationExternalData(application.externalData)
-    const iban = validIBAN(bankInfo.iban || '')
-
-    expect(false).toEqual(iban)
-  })
-
-  it('valid iban - should return true if the iban is right structured', () => {
-    const application = buildApplication({
-      externalData: {
-        socialInsuranceAdministrationApplicant: {
-          data: {
-            bankAccount: {
-              iban: 'NL91ABNA0417164300',
-              swift: 'NEDSZAJJXXX',
-              foreignBankName: 'Heiti banka',
-              foreignBankAddress: 'Heimili banka',
-              currency: 'EUR',
-            },
-          },
-          date: new Date(),
-          status: 'success',
-        },
-      },
-    })
-
-    const { bankInfo } = getApplicationExternalData(application.externalData)
-    const iban = validIBAN(bankInfo.iban || '')
-
-    expect(true).toEqual(iban)
-  })
-
-  it('valid swift - should return false because the country code should be letters', () => {
-    const application = buildApplication({
-      externalData: {
-        socialInsuranceAdministrationApplicant: {
-          data: {
-            bankAccount: {
-              iban: 'NLLLABNA0417164300',
-              swift: 'NE32ZAJJXXX',
-              foreignBankName: 'Heiti banka',
-              foreignBankAddress: 'Heimili banka',
-              currency: 'EUR',
-            },
-          },
-          date: new Date(),
-          status: 'success',
-        },
-      },
-    })
-
-    const { bankInfo } = getApplicationExternalData(application.externalData)
-    const swift = validSWIFT(bankInfo.swift || '')
-
-    expect(false).toEqual(swift)
-  })
-
-  it('valid swift - should return true if the swift is right structured', () => {
-    const application = buildApplication({
-      externalData: {
-        socialInsuranceAdministrationApplicant: {
-          data: {
-            bankAccount: {
-              iban: 'NL91ABNA0417164300',
-              swift: 'NEDSZAJJXXX',
-              foreignBankName: 'Heiti banka',
-              foreignBankAddress: 'Heimili banka',
-              currency: 'EUR',
-            },
-          },
-          date: new Date(),
-          status: 'success',
-        },
-      },
-    })
-
-    const { bankInfo } = getApplicationExternalData(application.externalData)
-    const swift = validSWIFT(bankInfo.swift || '')
-
-    expect(true).toEqual(swift)
   })
 })
