+++ libs/application/templates/social-insurance-administration/old-age-pension/src/types.ts
@@ -40,23 +40,6 @@ export interface ChildPensionRow {
   childDoesNotHaveNationalId: boolean
 }
 
-export interface BankInfo {
-  bank?: string
-  ledger?: string
-  accountNumber?: string
-  iban?: string
-  swift?: string
-  foreignBankName?: string
-  foreignBankAddress?: string
-  currency?: string
-}
-
-export interface ApplicantInfo {
-  bankAccount?: BankInfo
-  phoneNumber?: string
-  emailAddress?: string
-}
-
 export interface IncompleteEmployer {
   email?: string
   phoneNumber?: string
