+++ libs/clients/social-insurance-administration/src/clientConfig.yaml
@@ -1,7 +1,7 @@
 openapi: 3.0.0
 info:
   title: Social Insurance Administration API
-  version: 1.0.6
+  version: 1.0.7
   contact:
     name: Social Insurance Administration (Tryggingastofnun)
 servers:
@@ -36,13 +36,13 @@ paths:
           content:
             application/json:
               schema:
-                $ref: '#/components/schemas/OldAgePensionResponse'
+                $ref: '#/components/schemas/Response'
         '500':
           description: internal Error
           content:
             application/json:
               schema:
-                $ref: '#/components/schemas/OldAgePensionResponseError'
+                $ref: '#/components/schemas/ResponseError'
   /Applicant:
     get:
       tags:
@@ -214,12 +214,12 @@ components:
             - selfEmployed
         file:
           type: string
-    OldAgePensionResponse:
+    Response:
       type: object
       properties:
         applicationLineId:
           type: number
-    OldAgePensionResponseError:
+    ResponseError:
       type: object
       properties:
         code:
